uod minimap

+++ kana_v06/inc/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: avoronko <avoronko@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/07 10:17:30 by ksansom           #+#    #+#             */
-/*   Updated: 2024/05/31 12:47:39 by avoronko         ###   ########.fr       */
+/*   Updated: 2024/05/31 13:00:24 by avoronko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -94,12 +94,12 @@ typedef struct s_ray
 typedef struct s_wall
 {
 	int			orientation;
-	double		line_height;
-	double		draw_start;
-	double		draw_end;
+	int			line_height;
+	int			draw_start;
+	int			draw_end;
 	bool		vertical_wall;
-	double		tex_x;
-	double		tex_y;
+	int			tex_x;
+	int			tex_y;
 	double		wall_dist;
 	double		step;
 	double		tex_pos;
